- Remove all in-memory repository implementations and test files    - Remove NoopLogger (only needed for in-memory mode)    - Collapse service dual-path (createWithDBTx/createInMemory) into a      single straight-line db tx flow in both account and transaction services    - Config simplified to require postgres env vars unconditionally;      DBDriver constant and switch removed from main.go    - Add dedicated idempotency_keys table (key, request_hash, response_code,      response_body JSONB, created_at) instead of a UNIQUE column on transactions;      handler caches the full HTTP response and detects body conflicts via SHA-256    - Amount accepted as float64 rupees at HTTP boundary, stored as int64 paise      (×100 on input, ÷100 on output); DB column is BIGINT    - Merge 003/004 migrations into 001_init.sql; only two migration files remain    - Update README to document paise storage, idempotency design, and postgres-only setup

+++ internal/repository/transaction/repository.go
@@ -14,7 +14,7 @@ type Repository interface {
 	// tx must be a live transaction when the insert should be part of a larger
 	// atomic operation; pass nil to run outside any transaction.
 	// amount and txType are pre-computed by the service layer.
-	Create(ctx context.Context, tx *sql.Tx, accountID, operationTypeID int64, amount float64, txType string) (*model.Transaction, error)
+	Create(ctx context.Context, tx *sql.Tx, accountID, operationTypeID int64, amount int64, txType string) (*model.Transaction, error)
 
 	// FindOperationType retrieves an operation type by its numeric ID.
 	// Returns an error if the operation type does not exist.
